Added simple startscript that runs node as daemon, logging is verbose, to be controlled via options.
Implemented rss parsing and interaction with transmission, torrents not in ledger is added to transmission

A ledger is constantly updated with torrents that has been downloaded and added to transmission, the ledger contains the xt and timestap so old entries can be purged when not in the rss feed any more.

the rssget daemon polls the rss feed once every hour by default.